refactor(auth): remove session-based auth in favor of JWT
- Remove express-session and related middleware/configurations
- Consolidate auth controllers into single file
- Update env.example by removing SESSION_SECRET
- Update package.json and lock file dependencies
