+++ backend/controllers/auth.controller.js
@@ -2,7 +2,7 @@ import User from '../models/userSchema.js';
 import jwt from 'jsonwebtoken';
 import { JWT_EXPIRATION, COOKIE_EXPIRATION } from '../config/constants.js';
 
-// This function generates the token after a successful login
+// Generate JWT token and set cookie
 export const generateToken = (res, userId) => {
   const token = jwt.sign({ userId }, process.env.JWT_SECRET, {
     expiresIn: JWT_EXPIRATION,
@@ -16,17 +16,21 @@ export const generateToken = (res, userId) => {
   });
 };
 
-// This function logs the user out
-export const logout = (req, res) => {
-  res.cookie('jwt', '', {
-    httpOnly: true,
-    expires: new Date(0),
-  });
-  res.status(200).json({ message: 'Logged out successfully' });
-};
-
-// This is the callback function for GitHub OAuth
+// GitHub OAuth callback
 export const githubCallback = (req, res) => {
     generateToken(res, req.user._id);
-    res.redirect(process.env.FRONTEND_URL || '/');   // Redirect to the frontend application after successful login
+    res.redirect(`${process.env.FRONTEND_URL}/dashboard`);
+};
+
+// Get authenticated user profile
+export const profile = (req, res) => {
+    res.json({
+        username: req.user.username,
+    });
+};
+
+// Logout user
+export const logout = (req, res) => {
+    res.clearCookie("jwt");
+    res.json({ message: `logged out successfully` });
 };
