+++ backend/env.example
@@ -9,11 +9,8 @@ JWT_EXPIRES_IN=7d
 GITHUB_CLIENT_ID=Ov23liXO6S97Jh578vZS
 GITHUB_CLIENT_SECRET=8eae609cd26c7b22e9e468fe7142743950a7ce6e
 
-# ... existing variables ...
-
 # AI Configuration
 GEMINI_API_KEY=your_gemini_api_key_here
-SESSION_SECRET=your_session_secret_here
 
 # Frontend URL for redirects
 FRONTEND_URL=http://localhost:8080
